% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dummy_df.R
\name{generate_dummy_df}
\alias{generate_dummy_df}
\title{Generate dummy data frame}
\usage{
generate_dummy_df(
  rows_n = 10,
  cols_n = 5,
  col_types = c("numeric", "character", "factor", "range"),
  col_names = "var"
)
}
\arguments{
\item{rows_n}{integer value specifying the amount of rows for the generated data frame, default = 10}

\item{cols_n}{integer value specifying the amount of columns for the generated data frame, default = 5}

\item{col_types}{character string specifying the column types, cycles through all types by default or chosen types as such: \cr
'col_types = c("numeric", "factor")'. Available types are "numeric", "character", "factor" and "range"}

\item{col_names}{character vector specifying the used prefix for the col names, defaults to "var"}
}
\value{
data.frame
}
\description{
Generate dummy data frame
}
\examples{
set.seed(12345) # for reproducible results
example_dummy_df <- generate_dummy_df(rows_n = 12, cols_n = 6, col_types = c("character", "range"), col_names = "column")
}
