% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek_table.R
\name{peek_table}
\alias{peek_table}
\title{Create small tables for overview}
\usage{
peek_table(table_input, head_n = 6, tail_n = 6, ID = FALSE)
}
\arguments{
\item{table_input}{data.frame or tibble}

\item{head_n}{integer vector specifying row count for head}

\item{tail_n}{integer vector specifying row count for tail}

\item{ID}{TRUE by default, adds a row containing ID = nrow}
}
\value{
data.frame or tibble
}
\description{
Create small tables for overview
}
\examples{
example_table <- datasets::airquality
peek_table(example_table, head = 5, tail = 5)
}
