% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_to_numeric.R
\name{range_to_numeric}
\alias{range_to_numeric}
\title{Convert range in column to numeric}
\usage{
range_to_numeric(
  column,
  method = "middle",
  show_output = FALSE,
  round_result = NULL
)
}
\arguments{
\item{column}{\verb{<chr>} column containing data range in format "NUM-NUM" (e.g., \code{"32-34"})}

\item{method}{a character string specifying method by which to calculate resulting value from range. Prints minimum values (\code{"min"}),
maximum values (\code{"max"}), calculates average (\verb{"average'}), otherwise (default) (\code{"middle"}) is used.}

\item{show_output}{a logical specifying to print output to console, \code{FALSE} by default.}

\item{round_result}{an integer specifying the number of decimals (e.g., \code{"1"}), defaults to \code{NULL}; no rounding}
}
\value{
numeric vector.
}
\description{
Convert range in column to numeric
}
\examples{
# Example dataset:
data_table <- data.frame(
  column = c("28-29", "25-30", "34-38", "22-24")
)
data_table$column <-
  range_to_numeric(
    data_table$column,
    method = "middle",
    show_output = FALSE,
    round_result = 1)
}
