% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usable_colnames.R
\name{usable_colnames}
\alias{usable_colnames}
\title{Make column names usable}
\usage{
usable_colnames(input, exclude = "", space_to = "", case_lower = FALSE)
}
\arguments{
\item{input}{data.frame or tibble}

\item{exclude}{(string of) character(s) to exclude (e.g., " -_" results in colnames with spaces, hyphens and underscores intact)}

\item{space_to}{(string of) character(s) to change space to (e.g., "_" excludes space and converts it to underscores)}

\item{case_lower}{if TRUE converts all uppercase characters to lowercase. FALSE by default.}
}
\value{
character vector containing column names
}
\description{
Make column names usable
}
\note{
Currently has a bug when combining \code{exclude = "-"} with \code{space_to = "any input"}
}
\examples{
example_df <- data.frame(
 `test!@_option-25` = 1:5,
 `dATa &name` = 6:10,
 `column@#4$` = 11:15,
 check.names = FALSE
)

colnames(example_df) <-
usable_colnames(example_df,
  space_to = "_",
  case_lower = TRUE)
}
